+++ src/main/java/com/appsflyer/donkey/server/ServerImpl.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 AppsFlyer
+ * Copyright 2020-2021 AppsFlyer
  *
  * Licensed under the Apache License, Version 2.0 (the "License")
  * you may not use this file except in compliance with the License.
@@ -67,7 +67,9 @@ public final class ServerImpl implements Server {
       handlers.add(ServerHeaderHandler.create());
     }
   
-    handlers.forEach(h -> config.routeList().addFirst(RouteDefinition.create().handler(h)));
+    handlers.stream()
+            .map(h -> RouteDefinition.create().handler(h))
+            .forEach(rd -> config.routeList().addFirst(rd));
   }
   
   @Override